// apps/portal-api/src/model/kanel/public/ServiceGroup.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ServiceInstanceId } from './ServiceInstance';

/** Identifier type for public.ServiceGroup */
export type ServiceGroupId = string & { __brand: 'public.ServiceGroup' };

/** Represents the table public.ServiceGroup */
export default interface ServiceGroup {
  id: ServiceGroupId;

  name: string;

  service_instance_id: ServiceInstanceId;
}

/** Represents the initializer for the table public.ServiceGroup */
export interface ServiceGroupInitializer {
  /** Default value: gen_random_uuid() */
  id?: ServiceGroupId;

  name: string;

  service_instance_id: ServiceInstanceId;
}

/** Represents the mutator for the table public.ServiceGroup */
export interface ServiceGroupMutator {
  id?: ServiceGroupId;

  name?: string;

  service_instance_id?: ServiceInstanceId;
}